Added Http Servelet but want to update to Authetication

+++ src/main/java/com/example/dashbackend/services/FoodItemService.java
@@ -6,21 +6,23 @@ import com.example.dashbackend.entities.User;
 import com.example.dashbackend.repository.CategoryRepository;
 import com.example.dashbackend.repository.FoodItemRepository;
 import com.example.dashbackend.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class FoodItemService {
 
     private final CategoryRepository categoryRepository;
-    private final UserRepository userRepository;
     private final CategoryService categoryService;
     private final FoodItemRepository foodItemRepository;
 
-    public FoodItem createFoodItem(FoodItem foodItem, User user) {
+    public FoodItem createFoodItem(FoodItem foodItem, HttpServletRequest request) {
+        User user = categoryService.getUserFromRequest(request);
         Category existingCategory = categoryRepository.findById(foodItem.getCategory().getId())
                 .orElseThrow(() -> new RuntimeException("Category not found"));
 
@@ -29,19 +31,15 @@ public class FoodItemService {
         return foodItemRepository.save(foodItem);
     }
 
-    public List<Category> getUserCategories(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new RuntimeException("User not found"));
-        return categoryService.findByUserIdWithFoodItems(user.getId());
+    public FoodItem getFoodItemById(int id, HttpServletRequest request) {
+        User user = categoryService.getUserFromRequest(request);
+        Optional<FoodItem> foodItem = foodItemRepository.findById(id);
+        return foodItem.filter(f -> user.getId().equals(f.getCategory().getUser().getId())).orElseThrow(() -> new RuntimeException("FoodItem not found"));
     }
 
-    public FoodItem getFoodItemById(int id) {
-        return foodItemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("FoodItem not found"));
-    }
-
-    public FoodItem updateFoodItem(int id, FoodItem updatedFoodItem) {
-        FoodItem foodItem = foodItemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("FoodItem not found"));
+    public FoodItem updateFoodItem(int id, FoodItem updatedFoodItem, HttpServletRequest request) {
+        User user = categoryService.getUserFromRequest(request);
+        FoodItem foodItem = getFoodItemById(id, request);
 
         foodItem.setFoodName(updatedFoodItem.getFoodName());
         foodItem.setFoodDescription(updatedFoodItem.getFoodDescription());
@@ -57,7 +55,13 @@ public class FoodItemService {
         return foodItemRepository.save(foodItem);
     }
 
-    public void deleteFoodItem(int id) {
-        foodItemRepository.deleteById(id);
+    public void deleteFoodItem(int id, HttpServletRequest request) {
+        User user = categoryService.getUserFromRequest(request);
+        FoodItem foodItem = getFoodItemById(id, request);
+        if (user.getId().equals(foodItem.getCategory().getUser().getId())) {
+            foodItemRepository.deleteById(id);
+        }
     }
+
+
 }
